Use <select value> instead of <option selected>
